pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Akhil Gupta <akhilnidumukkula@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Jinja2 = "^2.11.2"
httpx = "^0.16.1"
black = "^20.8b1"
aiofiles = "^0.5.0"
attr = "^0.3.1"
attrs = "^20.2.0"
flake8 = "^3.8.4"
telnetlib3 = "^1.0.3"
netmiko = "^3.3.2"
"ruamel.yaml" = "^0.16.12"
ansible = "^2.10.3"
netdev = "^0.9.3"
uvloop = "^0.14.0"
nornir = "^3.0.0"
nornir-netmiko = "^0.1.1"
nornir-utils = "^0.1.1"
nornir-jinja2 = "^0.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
black = "^20.8b1"
mypy = "^0.790"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
